Refactor: Move Context.Provider to App
