+++ bridge-project/src/root.tsx
@@ -3,6 +3,7 @@ const Root = () => {
     <>
       <div className="container flex-row space-y-10">
         <h1>Bridge</h1>
+
         <nav>
           <ul className="space-y-5">
             <li>
